+++ Hotfix/Scene/Avatars/Handler/AvatarLeaveRoomHandler.cs
@@ -34,7 +34,7 @@ public sealed class AvatarLeaveRoomHandler : AddressRPC<FScene, AvatarLeaveRoomR
 
     private static long TryGetRoomId(InnerResult result)
     {
-        if (result.Args is { Count: > 0 } && result.Args[0] is long roomId)
+        if (result.Args is { Count: > 0 } && result.Args[0] is uint roomId)
         {
             return roomId;
         }
