+++ Hotfix/Scene/Rooms/Handler/RoomsLeaveHandler.cs
@@ -35,7 +35,7 @@ public sealed class RoomsLeaveHandler : AddressRPC<FScene, RoomsLeaveReq, RoomsL
 
     private static long TryGetRoomId(InnerResult result)
     {
-        if (result.Args is { Count: > 0 } && result.Args[0] is long roomId)
+        if (result.Args is { Count: > 0 } && result.Args[0] is uint roomId)
         {
             return roomId;
         }
